phase-3.1: extract RoundDataMap to its own file
RoundDataMap, RoundData, and HeightAndRound were ~140 lines buried in
the middle of state.go. They're a self-contained, already-thread-safe
data structure — they deserve their own file. Pure move, no behavior
change.
